+++ package.json
@@ -31,6 +31,7 @@
     "jest": "^21.2.1",
     "nps": "^5.7.1",
     "prettier": "^1.8.2",
+    "react-router-dom": "^4.2.2",
     "webpack": "^3.9.1",
     "webpack-dev-server": "^2.9.5",
     "webpack-node-externals": "^1.6.0"
