+++ src/App/index.js
@@ -1,90 +1,35 @@
-import React, { Component } from "react";
+import React from "react";
 import PropTypes from "prop-types";
+import { Route, Link } from "react-router-dom";
+import Home from "./screens/Home";
+import CrossTrainer from "./screens/CrossTrainer";
 
-import ScrambleHistory from "../ScrambleHistory";
+const App = ({ initialScramble }) => (
+  <div>
+    <nav>
+      <ul>
+        <li>
+          <Link to="/">Home</Link>
+        </li>
+        <li>
+          <Link to="/cross-trainer">Cross Trainer</Link>
+        </li>
+      </ul>
+    </nav>
 
-import getRandomSolve from "rubiks-cross-trainer";
-import createEmptyArray from "create-empty-array";
+    <hr />
 
-class App extends Component {
-  constructor(props) {
-    super(props);
-
-    this.optionsArray = createEmptyArray(8);
-    this.initialDifficulty = 1;
-
-    this.state = {
-      difficulty: this.initialDifficulty,
-      scramble: props.initialScramble || getRandomSolve(this.initialDifficulty),
-      scrambleHistory: []
-    };
-  }
-
-  getRandomSolve = () => {
-    this.setState(prevState => ({
-      scramble: getRandomSolve(prevState.difficulty),
-      scrambleHistory: [
-        {
-          difficulty: prevState.difficulty,
-          scramble: prevState.scramble
-        }
-      ].concat(prevState.scrambleHistory)
-    }));
-  };
-
-  handleSelectChange = e => {
-    this.setState({
-      difficulty: e.target.value
-    });
-  };
-
-  render() {
-    const containerStyle = {
-      display: "flex",
-      flexDirection: "column"
-    };
-
-    const headerStyle = {
-      textAlign: "center"
-    };
-
-    const formStyle = {
-      alignItems: "center",
-      display: "flex",
-      flexDirection: "column",
-      justifyContent: "center"
-    };
-
-    const selectStyle = {
-      marginBottom: "1rem"
-    };
-
-    return (
-      <div style={containerStyle}>
-        <h1 style={headerStyle}>{this.state.scramble}</h1>
-
-        <div style={formStyle}>
-          <div style={selectStyle}>
-            <span>Number of moves to generate cross:</span>
-            <select
-              value={this.state.difficulty}
-              onChange={this.handleSelectChange}
-            >
-              {this.optionsArray.map((_, index) => (
-                <option key={index} value={index + 1}>
-                  {index + 1}
-                </option>
-              ))}
-            </select>
-          </div>
-          <button onClick={this.getRandomSolve}>Get random solve</button>
-        </div>
-
-        <ScrambleHistory history={this.state.scrambleHistory} />
-      </div>
-    );
-  }
-}
+    <Route path="/" exact component={Home} />
+    <Route
+      path="/cross-trainer"
+      render={
+        /* istanbul ignore next */ () => (
+          <CrossTrainer initialScramble={initialScramble} />
+        )
+      }
+    />
+  </div>
+);
 
 App.propTypes = {
   initialScramble: PropTypes.string
