+++ src/App/screens/Home/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Home = () => <h1>Happy Cubing</h1>;
+
+export default Home;
